move basic card to components/
