+++ personal/lib/components/basic_card.dart
